+++ tests/unit/utility/tVersion.m
@@ -1,6 +1,14 @@
 classdef tVersion < matlab.unittest.TestCase
 % TVERSION Unit tests for "mag.version" function.
 
+    methods (TestMethodSetup)
+
+        % Make sure to clear the persistent state of "mag.version".
+        function clearFunctions(~)
+            clear("functions"); %#ok<CLFUNC>
+        end
+    end
+
     methods (Test)
 
         % Test that "mag.version" returns a version that is Semantic