+++ resources/ReleaseNotes.md
@@ -1,4 +1,4 @@
 # Software
 
 - Fix issue with determining version when toolbox is installed
-- Add tests for `mag.version` and `mag.buildtool.PackageTask`
+- Add tests for `mag.version`, `mag.buildtool.PackageTask` and `mag.graphics.chart.Scatterhistogram`